Added addObject to regions.py
The idea is to do a small paradigm shift.  Load annotation databases
into something like simple lists of objects.  Then, when overlap queries
are needed, simply build a RegionList with an attached reference to the
original object.  This keeps the original objects intact, etc., and building
the IntervalTrees for the RegionList is pretty fast, so there is not a need
to store it.  For example, dbSNP 130, with 18M rows can be loaded into
the RegionList in about 1.5 minutes.  Queries are then, of course, blazingly
fast and will return the attached object.